Added some projects

+++ Mst.java
@@ -2,7 +2,7 @@ import java.util.*;
 
 public interface Mst extends java.rmi.Remote
 {
-	int get_mst(String identifier)throws java.rmi.RemoteException;
+	long get_mst(String identifier)throws java.rmi.RemoteException;
 	void add_edge(String identifier,int from,int to,int weight)throws java.rmi.RemoteException;
 	void new_graph(String identifier,int n)throws java.rmi.RemoteException;
 	// void printGraph(Graph graph)throws java.rmi.RemoteException;
